Created Home page for student Dashboard so students are able to see the groups they are part of, and choose a group to view and rate teammates inside Fixed Code in App.tsx so that Student home is now Home.tsx in StudentDashboard In StudentDashboard.css, added .card:hover Made changes to saveRating to check if the rating was already saved. If so, updates the rating, if not saves new one. Remade previous Home.tsx to be the homepage for the Student Dashboard and created file Group.tsx to be where students can see the team members in a group they are in (basically old Home.tsx)
